+++ Software/ipc.py
@@ -6,7 +6,7 @@ every module imports from one authoritative place.
 Shared-memory layout (pre-allocated to maximum sizes to avoid runtime
 reallocation when parameters such as FFT_Length change):
 
-    SHM_SPECTRUM   (MAX_TOTAL_FFT,)                        float32
+    SHM_SPECTRUM   (MAX_TOTAL_FFT,)                        float16
     SHM_DETECTION  (512 x 512 x 3)                         uint8
 
 Active shapes are communicated via init_params; workers access only needed
@@ -29,10 +29,10 @@ MAX_WATERFALL_HEIGHT = 1024 # max waterfall history depth
 # ── Array specifications ───────────────────────────────────────────────────────
 
 SHM_SPECTRUM_SHAPE = (MAX_TOTAL_FFT,)  # (fft_pts,)
-SHM_SPECTRUM_DTYPE = np.float32
+SHM_SPECTRUM_DTYPE = np.float16
 
 SHM_WATERFALL_SHAPE = (MAX_WATERFALL_HEIGHT, MAX_TOTAL_FFT)
-SHM_WATERFALL_DTYPE = np.float32
+SHM_WATERFALL_DTYPE = np.float16
 
 SHM_DETECTION_SHAPE = (512, 512, 3)  # fixed UI display output
 SHM_DETECTION_DTYPE = np.uint8
